Cleanup nodes rendering calls
The nodes were cleared from one frame to another by redrawing them in
white. This does not work when the browser is in dark mode where the
background is dark because the canvas is transparent. This also has
issues with some border pixels remaining due to anti-aliasing and color
mixing maths.

Since all the nodes are redrawn every frame anyway, the canvas is
completly cleared first.