Simple matrix-based transformations
Nodes "world" coordinates are converted to "canvas" coordinates using
matrices. This greatly simplifies the math involved from our
perspective.

This fixes the previous rotation issues where all nodes were at the same
location and the rotation from the camera perspective (the canvas) was
not obvious as well.

With these changes the `Coordinates` class is redundant with the builtin
`DOMPoint` but was kept anyway.